% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DiffTestR_example_data_wide}
\alias{DiffTestR_example_data_wide}
\title{Example proteomics dataset (wide format)}
\format{
A data.table with 2,835 rows (precursors) and 11 columns:
\describe{
\item{Protein.Group}{Protein group identifier}
\item{Protein.Names}{Protein names with species suffix}
\item{Precursor.Id}{Unique precursor identifier}
\item{20230314_EXPL1_SA_Protoc_OT2_MixA_200ng_TR1_01}{Mix A replicate 1}
\item{20230314_EXPL1_SA_Protoc_OT2_MixA_200ng_TR2_01}{Mix A replicate 2}
\item{20230314_EXPL1_SA_Protoc_OT2_MixA_200ng_TR3_01}{Mix A replicate 3}
\item{20230314_EXPL1_SA_Protoc_OT2_MixA_200ng_TR4_01}{Mix A replicate 4}
\item{20230314_EXPL1_SA_Protoc_OT2_MixB_200ng_TR1_01}{Mix B replicate 1}
\item{20230314_EXPL1_SA_Protoc_OT2_MixB_200ng_TR2_01}{Mix B replicate 2}
\item{20230314_EXPL1_SA_Protoc_OT2_MixB_200ng_TR3_01}{Mix B replicate 3}
\item{20230314_EXPL1_SA_Protoc_OT2_MixB_200ng_TR4_01}{Mix B replicate 4}
}
}
\source{
Subset from HYE species mix benchmark (2023)
}
\usage{
DiffTestR_example_data_wide
}
\description{
Wide format version of the example dataset with precursor quantities
arranged in columns by MS run.
}
\details{
This is the wide format representation suitable for direct input to
\code{\link{testDifferentialAbundance}}.
}
\seealso{
\code{\link{DiffTestR_example_data_long}}, \code{\link{DiffTestR_example_study_design}}
}
\keyword{datasets}
