% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFromDIANN.R
\name{importFromDIANN}
\alias{importFromDIANN}
\title{Import precursor-level matrix from DIANN results}
\usage{
importFromDIANN(
  path_to_result = "../../04_DIANN_LfMS1/DIANN_LfMS1-Pass2.tsv",
  quant_value_col = "Precursor.Quantity",
  study_design = NULL
)
}
\arguments{
\item{path_to_result}{Path to DIANN result tsv file, long format}

\item{quant_value_col}{The name of the column containing precursor level quant
values to use. Typically Precursor.Quantity or Precursor.Normalised}

\item{study_design}{Study design in tab-separated .txt with mandatory columns filename, condition, replicate
By default, Study_design_template.tsv is written next to the input tsv. Fill in condition and replicate and save as
Study_design_filled.tsv. Then the next run of the function will likely go through ;-).}
}
\value{
A qdata result object (list) containing objects data, data_wide, matrix_raw, ann_col and ann_row
}
\description{
Import precursor-level matrix from DIANN results
}
\author{
Moritz Heusel
}
