% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddata_methods.R
\name{print.ddata}
\alias{print.ddata}
\title{Print method for ddata objects}
\usage{
\method{print}{ddata}(x, ...)
}
\arguments{
\item{x}{A ddata object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for ddata objects
}
