% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessDataQuality.R
\name{print.DiffTestR_QC}
\alias{print.DiffTestR_QC}
\title{Print method for DiffTestR_QC objects}
\usage{
\method{print}{DiffTestR_QC}(x, ...)
}
\arguments{
\item{x}{A DiffTestR_QC object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for DiffTestR_QC objects
}
