% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.multiDiffExpr}
\alias{print.multiDiffExpr}
\title{Print method for multiDiffExpr objects}
\usage{
\method{print}{multiDiffExpr}(x, ...)
}
\arguments{
\item{x}{A multiDiffExpr object (result from testMultipleComparisons)}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for multiDiffExpr objects
}
