% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata_methods.R
\name{print.qdata}
\alias{print.qdata}
\title{Print method for qdata objects}
\usage{
\method{print}{qdata}(x, ...)
}
\arguments{
\item{x}{A qdata object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for qdata objects
}
