% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata_methods.R
\name{qdata-methods}
\alias{qdata-methods}
\title{S3 Methods for qdata Objects}
\description{
S3 class methods for proteomics quantification data (qdata)
objects created by importFromDIANN() and importFromTables()
}
