% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddata_methods.R
\name{summary.ddata}
\alias{summary.ddata}
\title{Summary method for ddata objects}
\usage{
\method{summary}{ddata}(object, ...)
}
\arguments{
\item{object}{A ddata object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns summary statistics
}
\description{
Summary method for ddata objects
}
