% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{summary.diffExpr}
\alias{summary.diffExpr}
\title{Summary method for diffExpr objects}
\usage{
\method{summary}{diffExpr}(object, ...)
}
\arguments{
\item{object}{A diffExpr object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns summary statistics
}
\description{
Summary method for diffExpr objects
}
