% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata_methods.R
\name{summary.qdata}
\alias{summary.qdata}
\title{Summary method for qdata objects}
\usage{
\method{summary}{qdata}(object, ...)
}
\arguments{
\item{object}{A qdata object}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for qdata objects
}
