% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddata_methods.R
\name{validate_ddata}
\alias{validate_ddata}
\title{Validate ddata object structure}
\usage{
validate_ddata(x)
}
\arguments{
\item{x}{A ddata object}
}
\value{
TRUE if valid, stops with error if invalid
}
\description{
Validate ddata object structure
}
