% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata_methods.R
\name{validate_qdata}
\alias{validate_qdata}
\title{Validate qdata object structure}
\usage{
validate_qdata(x)
}
\arguments{
\item{x}{A qdata object}
}
\value{
TRUE if valid, stops with error if invalid
}
\description{
Validate qdata object structure
}
