% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFromTables.R
\name{importFromTables}
\alias{importFromTables}
\title{Import precursor-level matrix from .tsv table (wide format)}
\usage{
importFromTables(
  path_to_tsv = "../../04_DIANN_LfMS1/Table_precursor_wide.txt",
  precursor_annotation = "../../04_DIANN_LfMS1/Table_precursor_annotation.txt",
  study_design = "../../04_DIANN_LfMS1/Study_design.txt"
)
}
\arguments{
\item{path_to_tsv}{Path to wide format result tsv file, with column
names Protein.Group, Precursor.Id, filename1 filename2 etc}

\item{precursor_annotation}{Tsv format table with columns Precursor.Id,
Protein.Group and extended information required later on}

\item{study_design}{Study design in tab-separated .txt with mandatory
columns filename, condition, replicate}
}
\value{
A qdata result object (list) containing objects
data, data_wide, matrix_raw, ann_col, ann_row, study_design and source_file
}
\description{
Import precursor-level .tsv table (wide format)
}
\author{
Moritz Heusel
}
